All necessary functional for todo microservice

+++ todos/src/main/java/com/maksimpegov/todos/TodosService.java
@@ -1,35 +1,82 @@
 package com.maksimpegov.todos;
 
+import com.maksimpegov.todos.models.TodoServiceResponse;
 import com.maksimpegov.todos.models.Todo;
 import com.maksimpegov.todos.models.TodoRepository;
 import org.springframework.stereotype.Service;
 
-import java.lang.reflect.Array;
+import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 
 @Service
 public record TodosService(TodoRepository todoRepository) {
-    public List<Todo> getTodos(String userId) {
-//        List<Todo> todos = todoRepository.getAllByUserId(userId);
-//        return todos;
-        List<Todo> todos = List.of(new Todo("Sleep", "1", new java.util.Date()), new Todo("Eat", "2", new java.util.Date()));
-        return todos;
+
+    public TodoServiceResponse getTodos(String userId) {
+        try {
+            List<Todo> todos = todoRepository.getAllByUserId(userId);
+            return new TodoServiceResponse("200", "Success", todos);
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new TodoServiceResponse("404", "Something went wrong. " + e.getMessage(), null);
+        }
     }
 
-    public String addTodo(String userId, Todo todo) {
-        Todo todo1 = new Todo("Sleep", "1", new java.util.Date());
-        todoRepository.save(todo1);
-        return "OK";
+    public TodoServiceResponse addTodo(Todo todo) {
+        try {
+            todo.setCreatedAt(new Date());
+            todoRepository.save(todo);
+            return new TodoServiceResponse("201", "Success", Collections.singletonList(todo));
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new TodoServiceResponse("400", "Your todo is not valid. " + e.getMessage(), null);
+        }
     }
 
-    public String deleteTodo(String userId, String todoId) {
-        todoRepository.deleteById(Long.parseLong(todoId));
-        return "OK";
+    public TodoServiceResponse editTodo(Todo newTodoPart) {
+        try {
+            Todo oldTodo = todoRepository.getOne(newTodoPart.getId());
+            oldTodo.setUpdatedAt(new Date());
+
+            if (newTodoPart.getText() != null) {
+                oldTodo.setText(newTodoPart.getText());
+            } else if (newTodoPart.isCompleted()) {
+                oldTodo.setCompleted(true);
+                oldTodo.setClosedAt(new Date());
+            } else {
+                return new TodoServiceResponse("400", "Your request is not valid");
+            }
+
+            todoRepository.save(oldTodo);
+            return new TodoServiceResponse("200", "Success", Collections.singletonList(oldTodo));
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new TodoServiceResponse("404", "You may provide wrong todo id or todo. " + e.getMessage());
+        }
     }
 
-    public String clearTodos(String userId) {
-        Todo todo1 = new Todo("Sleep", "1", new java.util.Date());
-        todoRepository.deleteByUserId(userId);
-        return "OK";
+    public TodoServiceResponse deleteTodo(String todoId) {
+        try {
+            if (!todoRepository.existsById(Long.parseLong(todoId))) {
+                return new TodoServiceResponse("404", "Todo with this id does not exist.");
+            }
+
+            Todo deletedTodo = todoRepository.getOne(Long.parseLong(todoId));
+            todoRepository.deleteById(Long.parseLong(todoId));
+            return new TodoServiceResponse("200", "Success", Collections.singletonList(deletedTodo));
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new TodoServiceResponse("404", "Something went wrong. " + e.getMessage());
+        }
+    }
+
+    public TodoServiceResponse clearTodos(String userId) {
+        try {
+            todoRepository.deleteByUserId(userId);
+            return new TodoServiceResponse("200", "Success");
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new TodoServiceResponse("404", "Todo with this id does not exist. " + e.getMessage());
+        }
     }
 }
